refactor: 좋아요 갯수 추가

+++ backend/src/main/java/com/woowacourse/gongseek/like/application/LikeService.java
@@ -28,7 +28,6 @@ public class LikeService {
         Member member = getMember(appMember);
         Article article = getArticle(articleId);
 
-        article.addLikeCount();
         saveByExistsLike(member, article);
     }
 
@@ -51,6 +50,7 @@ public class LikeService {
     private void saveByExistsLike(Member member, Article article) {
         if (!likeRepository.existsByArticleIdAndMemberId(article.getId(), member.getId())) {
             likeRepository.save(new Like(article, member));
+            article.addLikeCount();
         }
     }
 
@@ -59,7 +59,13 @@ public class LikeService {
         Member member = getMember(appMember);
         Article article = getArticle(articleId);
 
-        likeRepository.deleteByArticleAndMember(article, member);
-        article.minusLikeCount();
+        deleteByExistsLike(member, article);
+    }
+
+    private void deleteByExistsLike(Member member, Article article) {
+        if (likeRepository.existsByArticleIdAndMemberId(article.getId(), member.getId())) {
+            likeRepository.deleteByArticleAndMember(article, member);
+            article.minusLikeCount();
+        }
     }
 }
